+++ src/app/Fake.DotNet.NuGet/NuGet.fs
@@ -1,6 +1,6 @@
 
 /// Contains helper functions and task which allow to inspect, create and publish [NuGet](https://www.nuget.org/) packages.
-/// There is also a tutorial about [nuget package creating](../create-nuget-package.html) available.
+/// There is also a tutorial about [nuget package creating](/dotnet-nuget.html) available.
 module Fake.DotNet.NuGet.NuGet
 
 open Fake.IO