+++ src/app/Fake.DotNet.Paket/ThisAssemblyInfo.fs
@@ -1,0 +1,6 @@
+namespace System
+open System.Runtime.CompilerServices
+
+[<assembly: InternalsVisibleTo("Fake.Core.UnitTests")>]
+[<assembly: InternalsVisibleTo("Fake.Core.IntegrationTests")>]
+do ()