+++ src/legacy/FakeLib/RestorePackageHelper.fs
@@ -1,6 +1,6 @@
 [<AutoOpen>]
 /// Contains tasks which allow to restore NuGet packages from a NuGet package feed like [nuget.org](http://www.nuget.org).
-/// There is also a tutorial about [nuget package restore](../nuget.html) available.
+/// There is also a tutorial about [nuget package restore](/nuget.html) available.
 [<System.Obsolete("Use Fake.DotNet.NuGet.Restore instead")>]
 module Fake.RestorePackageHelper